Seperate animations from utils
